+++ ci/test.sh
@@ -24,7 +24,8 @@ case "${CIRCLE_NODE_TOTAL:-1}" in
   unitTests
   e2eTests
   assembleBinaries
-  ci/linkerd-acceptance.sh
+  # FLAKEY #1226
+  # ci/linkerd-acceptance.sh
   ;;
 
 *)
@@ -39,7 +40,8 @@ case "${CIRCLE_NODE_TOTAL:-1}" in
     echo "Running e2e tests in worker ${worker}"
     e2eTests
     assembleBinaries
-    ci/linkerd-acceptance.sh
+    # FLAKEY #1226
+    #ci/linkerd-acceptance.sh
     ;;
 
   *)