+++ CHANGES.md
@@ -2,6 +2,11 @@
 
 * Upgrade to scala 2.12.
 * Fix connection leak when retrying on responses with chunked bodies.
+* Add io.l5d.http.allSuccessful and io.l5d.h2.allSuccessful response classifiers.
+* Fail on duplicate config file properties instead of silently taking the last
+  value.
+* Port numbers in k8s names will now have the service's port mapping applied.
+* Add `io.l5d.k8s.ns` namer.
 * Expire idle services and clients.
 
 ## 1.0.0 2017-04-24